old-tests/submission/Infraware/Forms/contents/Forms/Not Use/fieldset_disabled.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>FieldSet_disabled</h3>
	</p>
	<li>The fieldset control be disabled when the document was loaded.</li>
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
		<fieldset id="input_field" disabled>
 		</fieldset>
	</form>
	<script>

		var field = document.getElementById("input_field");

		if (typeof(field.disabled) == "boolean") {
			if (field.disabled != true) {
				test(function() {
					assert_true(false, "disabled attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("disabled attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
